booking-service/requirements.txt
sqlalchemy[asyncio]
asyncpg
python-dotenv
fastapi
uvicorn[standard]
httpx